Modularize the queries files Relates #62
