+++ server/controller/routes/admin/getUsers.js
@@ -1,4 +1,4 @@
-const getUsers = require('../../../database/queries/getUsers');
+const getUsers = require('../../../database/queries/users');
 
 const getUsersData = (req, res, next) => {
   getUsers()
