+++ server/controller/routes/admin/viewUsers.js
@@ -1,4 +1,4 @@
-const getUsersOfEvent = require('../../../database/queries/getUsersOfEvent');
+const getUsersOfEvent = require('../../../database/queries/users');
 
 
 const getUsersEvent = (req, res, next) => {
