+++ server/controller/routes/user/registeration.js
@@ -3,11 +3,11 @@ const nodemailer = require('nodemailer');
 const ical = require('ical-generator');
 const moment = require('moment');
 
-const checkUserExist = require('../../../database/queries/checkUserExist');
-const getEventDetalis = require('../../../database/queries/events.js/getEventDetalis');
-const getUsersCode = require('../../../database/queries/getUsersCode');
-const signUserAttend = require('../../../database/queries/signUserAttend');
-const alreadyBooked = require('../../../database/queries/events.js/checkAlreadyBooked');
+const checkUserExist = require('../../../database/queries/users');
+const getEventDetalis = require('../../../database/queries/events');
+const getUsersCode = require('../../../database/queries/users');
+const signUserAttend = require('../../../database/queries/users');
+const alreadyBooked = require('../../../database/queries/events');
 
 const checkUser = (req, res, next) => {
   const mobileRegExp = /^(\+?\d{0,4})?\s?-?\s?(\(?\d{3}\)?)\s?-?\s?(\(?\d{3}\)?)\s?-?\s?(\(?\d{4}\)?)?$/;
