+++ server/database/queries/admin/checkEmail.js
@@ -1,4 +1,4 @@
-const connection = require('../config/connection');
+const connection = require('../../config/connection');
 
 const checkEmail = (email) => connection
   .query('SELECT * FROM admin WHERE email=$1', [email]);
