+++ server/database/queries/admin/index.js
@@ -1,0 +1,7 @@
+const checkEmail = require('./checkEmail');
+const signAttendanceSql = require('./signAttendanceSql');
+
+module.exports = {
+  checkEmail,
+  signAttendanceSql,
+};
