+++ server/database/queries/events/checkAlreadyBooked.js
