+++ server/database/queries/events/getEventDetalis.js
