+++ server/database/queries/events/getEvents.js
