+++ server/database/queries/events/getEventsWithDate.js
