+++ server/database/queries/users/index.js
@@ -1,11 +1,22 @@
 const cancelPopUp = require('./cancelRegistratinQ');
 const checkUserExit = require('./checkUserExist');
 const getUser = require('./getUser');
+const getUsers = require('./getUsers');
+const getUsersCode = require('./getUsersCode');
+const getUsersOfEvent = require('./getUsersOfEvent');
+const newUser = require('./newUser');
+const signUserAttend = require('./signUserAttend');
+const userHasBooked = require('./userHasBooked');
 
 
 module.exports = {
   cancelPopUp,
   checkUserExit,
   getUser,
-
+  getUsers,
+  getUsersCode,
+  getUsersOfEvent,
+  newUser,
+  signUserAttend,
+  userHasBooked,
 };
